pyproject.toml
[build-system]
requires = ["maturin>=1.5,<2.0"]
build-backend = "maturin"

[project]
name = "glossapi"
version = "0.1.2"
description = "Academic document processing pipeline with Rust-powered markdown cleaning"
authors = [
    {name = "GlossAPI Team", email = "glossapi.team@eellak.gr"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    # Core pipeline deps
    "pandas>=1.3.0",
    "numpy<2",  # ORT+RapidOCR best compatibility
    "scikit-learn==1.6.1",
    "joblib>=1.0.0",
    "dask>=2022.1.0",
    "pyarrow>=7.0.0",
    "aiohttp>=3.8.0",
    "aiofiles>=23.0.0",
    "ftfy>=6.0.0",
    "tenacity>=8.0.0",
    "tqdm>=4.67.0",
    "pyyaml>=6.0",
    # Math/JSON enrichment helpers
    "pypdfium2>=4.0.0",
    "zstandard>=0.22.0",
]
license = "EUPL-1.2"
classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Rust",
    "License :: OSI Approved :: European Union Public Licence 1.2 (EUPL 1.2)",
]

[project.optional-dependencies]
# Docling + RapidOCR ONNX stack (kept optional to preserve import-light installs)
rapidocr = [
    "docling==2.48.0",
    # Use RapidOCR core package; avoid rapidocr_onnxruntime to prevent pip
    # from auto-installing the CPU-only 'onnxruntime' wheel.
    "rapidocr>=3.3.0",
    "onnxruntime-gpu==1.18.1",
]
# Optional CUDA layout acceleration (Docling)
cuda = [
    "torch==2.5.1",
    "torchvision==0.20.1",
]
# DeepSeek OCR backend extras (CUDA 12.1 build of vLLM). Torch is not pinned here
# because users should install the CUDA wheel from the PyTorch index
# (see docs: installing torch==2.5.1+cu121 via extra index URL).
deepseek = [
    "vllm>=0.11.0",
    "transformers>=4.45,<5",
    "accelerate>=1.2.1,<2",
    "pymupdf==1.24.10",
    "Pillow==10.4.0",
]
docs = [
    "mkdocs>=1.5",
    "mkdocs-material>=9.5",
]

[tool.maturin]
bindings = "pyo3"
# The crate is located under rust/glossapi_rs_noise
module-name = "glossapi_rs_noise"
python-source = "src"

manifest-path = "rust/glossapi_rs_noise/Cargo.toml"
include = ["src/**"]
python-packages = [
    "glossapi"
]

[tool.pytest.ini_options]
markers = [
    "rapidocr: requires the RapidOCR/Docling execution stack",
    "deepseek: exercises the DeepSeek OCR pipeline",
]
